+++ vispy/mpl_plot/__init__.py
@@ -2,6 +2,16 @@
 # Copyright (c) 2014, Vispy Development Team.
 # Distributed under the (new) BSD License. See LICENSE.txt for more info.
 
+"""
+Matplotlib plotting backend wrapper.
+
+This module enables converting matplotlib plotting commands to
+vispy plotting commands. Support is experimental and incomplete,
+proceed with caution.
+"""
+
+__all__ = ['show']
+
 try:
     from matplotlib.pyplot import *  # noqa
 except ImportError: